feat: update order status when reverse inventory

+++ inventory-service/src/main/java/com/nduyhai/inventory/domain/ReservedStock.java
@@ -2,4 +2,4 @@ package com.nduyhai.inventory.domain;
 
 import java.util.UUID;
 
-public record ReservedStock(UUID productId, int quantity) {}
+public record ReservedStock(UUID productId, long quantity) {}
